+++ src/routes/[[lang]]/(auth)/login/(methods)/questions/+page.svelte
@@ -13,8 +13,8 @@
 	import { checkKeypairs, generateDid, saveUserPublicKeys, userEmailStore } from '../../_lib';
 	import type { Feedback } from '$lib/utils/types';
 	import { log } from '$lib/log';
-	import Loading from '$lib/components/molecules/Loading.svelte';
-	import { routeHistory } from '$lib/routeStore';
+	import FingerPrint from '$lib/assets/lottieFingerPrint/FingerPrint.svelte';
+	import HeaderWithBackButton  from '$lib/components/molecules/HeaderWithBackButton.svelte';
 
 	//
 
@@ -74,7 +74,6 @@
 
 				seed = keypair.seed;
 				loading = false;
-				
 			} catch (e) {
 				loading = false;
 				feedback = {
@@ -101,27 +100,27 @@
 	}
 
 	const goToWallet = () => {
-		goto('/wallet', undefined, false).then(() => {
-			routeHistory.clear();
-		});
+		goto('/wallet', undefined);
 	};
 </script>
 
-<Loading {loading} message={m.Generating_Keypair_()} />
-<d-header back-button={!seed} backFunction={routeHistory.back}>
+<d-loading {loading} message={m.Generating_Keypair_()}>
+	<FingerPrint />
+</d-loading>
+<HeaderWithBackButton>
 	{m.SECURITY_QUESTIONS()}
-</d-header>
+</HeaderWithBackButton>
 
 <div class="flex h-full w-screen flex-col gap-4 px-4">
 	<d-feedback {...feedback} />
 
 	{#if !seed}
-		<div class="flex flex-col gap-2">
+		<d-list>
 			<d-heading sixe="s">{m.Answer_to_these_questions()}</d-heading>
 			<d-text size="l"
 				>{m.to_ensure_the_security_of_your_account_and_simplify_key_recovery_please_answer_the_following_questions_()}</d-text
 			>
-		</div>
+		</d-list>
 
 		<Form
 			{form}
@@ -163,11 +162,11 @@
 		</d-button>
 	{:else}
 		<div class="flex h-screen flex-col place-content-between">
-			<div>
-				<div class="flex flex-col gap-2 pt-11">
+			<div class="pt-11">
+				<d-list>
 					<d-heading sixe="s">{m.Store_this_keypair()}</d-heading>
 					<d-text size="l">{m.your_unique_keypair_has_been_generated_successfully_()}</d-text>
-				</div>
+				</d-list>
 
 				<div class="flex w-full flex-col space-y-8 pb-6 pt-4">
 					<div class="flex flex-col gap-6">